fix[install_standard.sh]: Add check for nvim file

+++ .bin/install_standard.sh
@@ -22,7 +22,9 @@ curl -sL https://deb.nodesource.com/setup_13.x | sudo -E bash -
 sudo apt-get install -y nodejs
 
 # Install neovim
-wget "https://github.com/neovim/neovim/releases/download/v0.4.3/nvim.appimage" -O "$HOME/.bin/nvim.appimage" && chmod u+x "$HOME/.bin/nvim.appimage"
+if [ ! -f "$HOME/.bin/nvim.appimage" ]; then
+	wget "https://github.com/neovim/neovim/releases/download/v0.4.3/nvim.appimage" -O "$HOME/.bin/nvim.appimage" && chmod u+x "$HOME/.bin/nvim.appimage"
+fi
 pip install neovim
 sudo npm install -g neovim
